requirements.txt
ultralytics
opencv-python-headless
pillow
torch
torchvision
gdown
flask
flask-cors
gunicorn
